# docs/api-specification.yaml
...
openapi: 3.0.3
info:
  title: RocketVault API
  description: |
    A self-hosted, open-source alternative to Azure Key Vault: secrets
    management, cryptographic key operations, X.509 certificate lifecycle
    management, and multi-vault RBAC.
  version: 4.0.0
  contact:
    name: RocketVault
    url: https://github.com/Snehal1112/rocketvault
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.rocketvault.local
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/oauth2/token:
    post:
      summary: Issue a service-account access token
      description: |
        RFC 6749 §4.4 client-credentials grant. Credentials may be supplied via
        HTTP Basic authentication or as form fields in the request body. The
        body must be `application/x-www-form-urlencoded`. Parameters in the URL
        query string, including `client_secret`, are ignored.
      operationId: issueOAuth2Token
      tags:
        - Service Accounts
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/OAuth2TokenRequest"
      responses:
        "200":
          description: Access token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2TokenResponse"
        "400":
          description: Malformed request or unsupported grant type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2ErrorResponse"
        "401":
          description: Client authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2ErrorResponse"
        "500":
          description: Authorization server temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2ErrorResponse"
  /api/v1/service-accounts:
    post:
      summary: Create a service account
      description: |
        Creates an OAuth2 client-credentials identity and returns the plain-text
        client secret once only. Admin role required.
      operationId: createServiceAccount
      tags:
        - Service Accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceAccountRequest"
      responses:
        "201":
          description: Service account created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccountWithSecret"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      summary: List service accounts
      description: Returns all service accounts. Admin role required.
      operationId: listServiceAccounts
      tags:
        - Service Accounts
      responses:
        "200":
          description: Service accounts retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListServiceAccountsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/service-accounts/{service_account_id}:
    parameters:
      - $ref: "#/components/parameters/ServiceAccountId"
    get:
      summary: Get a service account
      description: Returns a single service account by ID. Admin role required.
      operationId: getServiceAccount
      tags:
        - Service Accounts
      responses:
        "200":
          description: Service account retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccount"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a service account
      description: Permanently deletes a service account. Admin role required.
      operationId: deleteServiceAccount
      tags:
        - Service Accounts
      responses:
        "200":
          $ref: "#/components/responses/StatusOK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/service-accounts/{service_account_id}/rotate:
    parameters:
      - $ref: "#/components/parameters/ServiceAccountId"
    post:
      summary: Rotate a service account's client secret
      description: |
        Issues a new client secret and invalidates the old one. The new secret
        is returned once only. Admin role required.
      operationId: rotateServiceAccountSecret
      tags:
        - Service Accounts
      responses:
        "200":
          description: Client secret rotated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientSecretResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─── OIDC ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `POST /api/v1/users/login`,
        `POST /api/v1/users/refresh`, the OIDC callback, or
        `POST /api/v1/oauth2/token`. Tokens are signed asymmetrically
        (RS256/ES256) and always carry a `kid` header.
  parameters:
...
    ServiceAccountId:
      name: service_account_id
      in: path
      required: true
      description: Unique identifier of the service account
      schema:
        type: string
        format: uuid
...
  responses:
    StatusOK:
      description: Operation completed successfully
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StatusResponse"
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or the session is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The caller is not authorized for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - id
        - message
        - status_code
      properties:
        id:
          type: string
          description: Error identifier
          example: "api.secret.create"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request body"
        detailed_error:
          type: string
          description: >-
            Either empty or a fixed, non-diagnostic string. Internal errors
            never echo driver, SQL or provider text; the generic 500 detail
            tells the caller to quote the request_id, which correlates the
            response with the server logs.
          example: "An internal error occurred. Quote the request_id when contacting support."
        status_code:
          type: integer
          description: HTTP status code
          example: 400
...
    StatusResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: "OK"
...
    OAuth2TokenRequest:
      type: object
      required:
        - grant_type
      properties:
        grant_type:
          type: string
          enum: [client_credentials]
        client_id:
          type: string
          description: Omitted when credentials are sent via HTTP Basic
        client_secret:
          type: string
          description: Omitted when credentials are sent via HTTP Basic
    OAuth2TokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: "Bearer"
        expires_in:
          type: integer
          description: Token lifetime in seconds
          example: 900
    OAuth2ErrorResponse:
      type: object
      required:
        - error
        - error_description
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - unsupported_grant_type
            - server_error
        error_description:
          type: string
    CreateServiceAccountRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "ci-pipeline"
        description:
          type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
    ServiceAccount:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
    ServiceAccountWithSecret:
      allOf:
        - $ref: "#/components/schemas/ServiceAccount"
        - type: object
          required:
            - client_secret
          properties:
            client_secret:
              type: string
              description: Plain-text client secret; returned once only
    ListServiceAccountsResponse:
      type: object
      required:
        - service_accounts
        - total
      properties:
        service_accounts:
          type: array
          items:
            $ref: "#/components/schemas/ServiceAccount"
        total:
          type: integer
    ClientSecretResponse:
      type: object
      required:
        - client_secret
      properties:
        client_secret:
          type: string
          description: New plain-text client secret; returned once only
...